% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bc.R
\name{bc}
\alias{bc}
\alias{bc_inv}
\alias{bc_prime}
\title{Box-Cox transformation}
\usage{
bc(y, lambda)

bc_inv(y, lambda)

bc_prime(y, lambda)
}
\arguments{
\item{y}{Numeric vector (must be >= 0).}

\item{lambda}{Numeric scalar, power parameter.}
}
\value{
\code{bc()}: Returns the transform value(s).

\code{bc_inv()}: Computes the inverse of \code{bc()}.

\code{bc_prime()}: Gives the derivative of \code{bc()}
  with respect to \code{y}.
}
\description{
These functions calculate the Box-Cox transformation, its inverse, and derivative.
}
\section{Functions}{
\itemize{
\item \code{bc}: The Box-Cox transformation

\item \code{bc_inv}: Inverse of the Box-Cox transformation

\item \code{bc_prime}: Derivative of the Box-Cox transformation
}}

\examples{
bc(0.5, 1.5)
bc(1, 0)

bc_inv(-0.4, 1.5)
bc_inv(0, 0)

bc_prime(0.5, 1.5)
bc_prime(1, 0)

}
