% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sdemodel_display.R
\name{sdemodel_display}
\alias{sdemodel_display}
\title{Display model specification}
\usage{
sdemodel_display(model)
}
\arguments{
\item{model}{SDE model specification, as produced by sdemodel()}
}
\value{
Invisibly returns its argument
}
\description{
Display model specification
}
\examples{
  mod <- sdemodel(); sdemodel_display(mod)
}
