% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/str2fun_theta.R
\name{str2fun_theta}
\alias{str2fun_theta}
\title{String to function, with parameters in theta}
\usage{
str2fun_theta(s)
}
\arguments{
\item{s}{String representation of a function of \code{x} and parameters}
}
\value{
Function of \code{x} and \code{theta}, \code{theta} being a named vector or list of parameters.
}
\description{
Normally not called by the user directly, used by \code{\link{sdefit}()}.
  Converts an expression, in a character string, to a function.
}
\examples{
str2fun_theta("x^c / a")
}
