% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/phi.R
\name{userphi}
\alias{userphi}
\alias{userphiprime}
\title{Optional external transformation and derivative functions}
\usage{
userphi(x, theta)

userphiprime(x, theta)
}
\arguments{
\item{x}{Numeric vector, variable to be transformed.}

\item{theta}{Named list of transformation parameters}
}
\value{
:  Transformed variable

:  Transformation derivative
}
\description{
Skeletons for user-supplied transformation and drivative functions, used by
  \code{sdefit()} if specified in parameters \code{phi} and/or
  \code{phiprime}. To be completed by the user.
}
\section{Functions}{
\itemize{
\item \code{userphi()}: Transformation

\item \code{userphiprime()}: Derivative

}}
\examples{
\dontrun{   userphi(20, list(a=70, c=0.5))   }
\dontrun{   userphiprime(20, list(a=70, c=0.5))   }
}
\keyword{internal}
