% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/phi.R
\name{phi}
\alias{phi}
\alias{phiprime}
\title{External transformation and derivative functions}
\usage{
phi(x, theta)

phiprime(x, theta)
}
\arguments{
\item{x}{Numeric vector, variable to be transformed.}

\item{theta}{Named list of transformation parameters}
}
\value{
\code{phi}:  Transformed variable

\code{phiprime}:  Transformation derivative
}
\description{
Dummy external transformation definitions, called by \code{sdefit()}
  if \code{trfuns == TRUE}. To be completed by the user.
}
\section{Functions}{
\itemize{
\item \code{phi}: Transformation

\item \code{phiprime}: Derivative
}}

\examples{
\dontrun{   phi(20, list(a=70, c=0.5))   }
\dontrun{   phiprime(20, list(a=70, c=0.5))   }
}
\keyword{internal}
